Add a new jenkins job to update tests-for.json.
Summary:
This is a json file that maps each file in webapp to the python and
javascript tests that access that file.  We can use this to limit what
tests we need to run when a file changes.

Computing the tests-for mapping is quite expensive, so we do it
periodically in a jenkins job, and check in the results so people can
use it easily.

Test Plan: Fingers super-crossed.  I'm sure this won't work the first time.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D33623