Fix the script to not pass through 1024 bytes from sys.stdin.
My logic for when to append to `input` was wrong.  It was only doing
it the first time we read in a chunk of data, not the rest of the
times, because it was only triggering when p.stdin was not alrady in
the write_set.  My brain was just fuzzy.

Test plan:
   cat /etc/ca-certificates.conf | ../jenkins-tools/timeout_output.py 45m sat | wc -c
Before this change it said 1024, after this change it said 9112, which
is the entire file.

Auditors: amos