Don't bother trying to run testutil.lint_test.
That is the "test" that runs all the linters, but we already know what
dependencies each linter has: by definition, a linter runs only over
the tests that are passed in to it.  So we can save some time by not
bothering to run them.

Auditors: benkraft

Test plan:
None