Add resolveCommitish from D33517.
I'm not quite ready to land D33517, but resolving the commit early is
something that it turns out we want to do for our existing tests as
well.  So I extract that one piece of code out, and use it.

Auditors: amos, benkraft, colin, chris

Test plan:
Will try running determine-test-dependencies after this is in.