Make sure we fail if any of the build.lib commands fail.
build.lib does not set `-e` explicitly.  In old jenkins scripts that
didn't matter because we sourced the file (using `.`) so it inherited
the `-e` from its parent.  But now we run it as a script instead of
sourcing it, so we need to run it under `-e` explicitly.

I could do `set -e` in the script itself but that was a bit more
invasive a change than I felt comfortable with.  In groovy-land, all
uses of build.lib go through kaGit, so it should be fine to just do
things here.  That also lets me specify `-x`, which while verbose I
think will be useful for debugging.

Auditors: amos, chris, colin

Test plan:
Fingers crossed