I was neglecting to delete the old json files before creating the new!
As a result, worker machines were stashing the latest results but also
results from previous runs, and when we copied them over to master, it
was a mix of old runs and current ones.  That's why we were seeing
failures due to "not all json files have the same commit."

Auditors: benkraft

Test plan:
Pretty confident about this one.