Another attempt to get cron working again.
The problem turned out to be -- after much debugging -- that I was
naming my method with the magic java `setFoo` naming convention, which
means that calling it as a function did weird things.  Just renaming
it to addCronSchedule fixes all my problems.

Auditors: amos, chris, colin, benkraft

Test plan:
I ran this on jenkins before committing (using the `replay` feature of
jenkins pipeline jobs), but the proof will be in the
running-from-scratch.