Fix a bunch of errors in exec.
1) I was forgetting to return the value from my function.
2) I was calling `sh` wrong: you need the `script` arg when also
   passing in returnStdout and returnStatus.  I knew that, but...
3) I was inadvertently creating a global variable by not putting a
   `def` in front of the `i`.

Normally groovy would catch (3) at least, but the way pipeline works
with files in `vars`, it creates globals for you willy-nilly, since it
thinks that you //want// to create a global var if you're putting it
in the vars/ directory.

Auditors: amos, chris, colin

Test Plan:
I ran e2e-test on jenkins with this new setup and it at least got to
running the tests.