Initialize globals within the notify() context.
Before this change, if global-initialization failed -- say because
params.NUM_WORKER_MACHINES was set to something that's not a number --
the job would fail but we would get no notification of that failure in
slack, because the global-initialization was happening outside the
notify() context.

Now I do the initialization inside notify(), so we do get such
notifications.

Auditors: amos, colin, chris

Test Plan:
Will run on e2e-test on jenkins.  The first time, I will set
NUM_WORKER_MACHINES to `a` and make sure we are notified of the
failure.