Add support for BACK TO NORMAL.
Now you can say "only send to slack when this job is back to normal"
-- that is, it succeeded after having failed in the previous run.

I refactored the code a bit to support BACK TO NORMAL, which sometimes
is a synonym for SUCCESS and sometimes isn't.

Auditors: amos, colin, chris

Test Plan:
This ran without error:
   groovy vars/notify.groovy

If I did cause a problem it will evidence itself after the body of the
script has run, so maybe the notification will be off but otherwise no
harm done.  So I feel OK just pushing this.