A script to help convert old-style jenkins jobs to groovy jobs.
Summary:
We prefer groovy jenkins jobs to freestyle jenkins jobs both because
they're easier to reason about, being written in an actual programming
language rather than a clunky web UI, and because we can control the
jenkins jobs via source control just like everything else we do.  (In
particular, we can make sure jenkins changes are code reviewed first,
which is difficult to do when using the web UI.)

This is a 'throwaway' script in that it won't be needed once
everything is converted.  So I didn't make much effort to handle
errors gracefully.  The output will probably need to be munged to be
style-clean.

Test Plan:
Ran
   ./convert_params.py ../jenkins/jobs/i18n-download-translations/config.xml
and the output looked reasonable.

Reviewers: colin, amos, chris

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D33795