Print a message to make it easier to identify what caused a job to fail.
After a job fails, we (may) execute a whole bunch of code to send
messages to hipchat, etc.  That code shows up at the end of the
console log, making it difficult to find the actual cause of the
error.

By adding in this red text, it should be easier to scroll up to the
place where the job is actually failing.

Auditors: amos, chris, colin

Test plan:
Fingers crossed.  If it fails, it will fail after the job has run, and
thus will hopefully not cause too much trouble.