First attempt at a groovy job for running webapp tests.
Summary:
This code is based on the existing groovy job for running e2e-tests.
(In fact, there's a lot of duplicated or almost-duplicated code, that
I may try to extract out into shared libraries one day.)

The main difference is that I kept the 'singleton' semantics of the
existing make-check/etc jobs: if we pass tests at a given commit, then
the next time we run tests at that commit it's a noop.  In theory,
this can save time if we end up having to re-deploy something due to a
gae failure.

This combines make-check, make-check-static, make-check-dynamic, and
make-check-worker into one rule: yay!  It also could run
make-allcheck, but doesn't have the cron support that make-allcheck
needs, so I'll have to write a small driver job for that.

Add a script for `make-allcheck`.

This one is really simple: it just calls webapp-test and e2e-test with
the appropriate parameters.

Test Plan:
This is entirely untested -- once it's been reviewed I'll actually
create a job on jenkins for it and see how far it goes.

We'll see!

Reviewers: amos, colin, chris, benkraft

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D33517