Make a new sanity check now that we are using the all.pot.pickle file.
Summary:
We switched to using the all.pot.pickle file rather than a debug txt
file as that file itself had a bug in it at some point and is hence
probably not as safe to use going forward.  That however means it is
harder to grep something in it.

Test Plan:
Put the following into test.sh
```
 #!/bin/bash -xe
ALL_POT="$PWD"/genfiles/translations/all.pot.pickle
echo -e "import sys, cPickle, polib\nwith open('$ALL_POT') as f:\n data=cPickle.load(f)\nif not isinstance(data, polib.POFile) or len(data) < 100000:\n  sys.exit(1)" | python
```
ran it and saw with 'echo $?' that it exited with 0.  Switched < to > and saw it exit with 1.

Reviewers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D33840