Modify _numConsecutiveFailures to take current status as an input.
Looking at currentBuild.result is a fool's game, because it is
typically null while the script is still executing.  We could manually
set it ourselves earlier, but an easier fix is to just pass in the
status of the current job, like we do elsewhere in this file.

Auditors: amos, chris, colin

Test plan:
Will try it out on a job that's been failing a lot.