Make sure that gitTag is actually a valid git tag.
We were using `current_version.py --git-tag` which does not produce a
current git tag when we're in the middle of a deploy -- in particular
when we've deployed a new version to GAE, but not yet pushed the tag
associated with it to git.  This actually is the state quite often,
since we have 5-10 minutes of monitoring between gae-deploy and
deploy-finish.

Auditors: james

Test plan:
I ran
   groovy -e "println 'foo\nbar\nbaz\nbang\n'.split('\n')[-1]"
and it said `bang`.  Good enough for me.