Add a groovy script for the content-publish e2e test.
Summary:
This is similar in effect to the regular e2e test, but much different
in implementation: it uses saucelabs exclusively instead of our local
workers, and runs on a timer rather than as part of the deploy
process.  So surprisingly little code is shared.

Test Plan:
Fingers crossed.  I ran
   groovy jobs/content-publish-e2e-test.groovy
successfully though, so I'm avoiding obvious typos.

Reviewers: amos, benkraft

Reviewed By: amos

Subscribers: chris, colin

Differential Revision: https://phabricator.khanacademy.org/D33877