Don't try to use default args for exec().
My understanding of default args in groovy was very flawed.  I thought
it worked like python, where you could say:
   def a(b, d=5, e=6) { ... }
and then call it like
   a("hello", e=7);

But while that's syntactically legal, it's not how things work at
all.  (Apparently the latter assigns b to "hello", but then assigns
*d* to 7, and leaves e as 6, or maybe null?  Ugh.)

But that's now how the default builtins work in groovy.  They do
something where a function either takes one argument, or it takes a
dict (like **kwargs in python).  Since `exec` is meant to be a drop-in
replacement for `sh`, I tried to do the same thing -- but ran into an
error where the pipeline script complained it could not serialize the
hash.

There are probably ways to resolve all this, but for now I just get
rid of the default args for exec.  If you want to use them, you'll
need to use sh and call escapeArgList explicitly.  We only do that in
one place so far.

Auditors: amos, chris, colin

Test Plan:
Fingers crossed.