Work around a bug in getLogMatcher() when result is null.
I don't know how this works in freestyle-project-land, but it seems to
be a pretty big flaw for something that can be called from a groovy
script at any time.  Luckily for us it's not hard to work around.

Auditors: amos, chris, colin

Test plan:
Will run i18n-update-string and then abort it, and see if it sends out mail.