Don't try to set the status back to null -- that's not allowed.
This is sub-optimal because our idea of the status may be wrong, and
once we do this assignment jenkins won't fix it.  This probably means
we'll never report 'ABORTED'.

Maybe I'll report the jenkins bug so if this is fixed we can remove
this workaround.

Auditors: amos, chris, colin

Test plan:
Will run i18n-update-string and then abort it, and see if it sends out mail.