Use rawBuild to get the logfile.
This has the advantage that a) it works, and b) it avoids the bug in
getLogMatcher() when the result is still null.  It has the
disadvantage that using rawBuild is a bit of a security hole.  I've
let it go for now, but I've marked a TODO.

Auditors: amos, colin, chris, james

Test plan:
Ran with this code on i18n-update-strings and it sent mail with
loglines!