Move the main boddy of the build-current-sqlite jenkins job to a shell script.
Summary:
This is following the lead of other jobs like i18n-gcs-upload and
sync-captions.  It is also much easier, in a groovy world, to call
out to a shell script than have a ton of `sh` commands inside groovy.

This isn't used yet, but once it's checked in I'll change the existing
build-current-sqlite job to just call out to it.

My only changes were to add double-quotes in lots of places, and to
change the formatting a bit (4-space indents, etc).

Test Plan: Fingers crossed.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D33904