Oops, I forgot that `a..b` includes b, so we need `end` to be size()-1.
This only matters when sending mail when a job succeeds.  We needed to
do that for the first time!

Auditors: amos, chris, colin

Test plan:
I ran the following code without error:
   groovy -e 'loglines = ["a", "c", "b"]; def end = loglines.size() - 1; for (def i = 0; i < loglines.size(); i++) { if ("===== JOB FAILED =====" in loglines[i]) { end = i; break; } }; def NUM_LINES = 50; def start = end - NUM_LINES < 1 ? 1 : end - NUM_LINES; println loglines[0, start..end].join("\n");'
It printed `a\nc\nb` as expected.  (This code basically matches my
changed code.)

I also did
   groovy -e 'loglines = ["a", "c", "b"] * 100; def end = loglines.size() - 1; for (def i = 0; i < loglines.size(); i++) { if ("===== JOB FAILED =====" in loglines[i]) { end = i; break; } }; def NUM_LINES = 50; def start = end - NUM_LINES < 1 ? 1 : end - NUM_LINES; println loglines[0, start..end].join("\n");'
and it printed 52 lines.  So I have an off-by-one error somewhere, but
I kinda suspected that.