Add groovy scripts for the znd-related jobs.
Summary:
There are two jobs: one to notify znd owners, and one to delete znd's.

I took this opportunity to finally implement the extraText argument to
notify; I thought of a way to get the right data to the right place at
the right time using closures (and, sadly, global variables).  We'll
see if it works!

Test Plan:
I ran:
   groovy jobs/delete-znd.groovy
   groovy jobs/notify-znd-owners.groovy

But the real test will be to run the jobs once they're landed.  I'll
run notify-znd-owners for real (but let people know first), and I'll
try delete-znd with error args (to test extraText) and then with a
real arg if anyone happens to answer saying I can delet theirs.

Reviewers: sean

Reviewed By: sean

Subscribers: colin, chris, amos

Differential Revision: https://phabricator.khanacademy.org/D33965