Groovy version of the sync-captions job.
Summary:
This job does everything in a single shell script, so it was easy
enough to move over!

I changed it to tell slack on success.  I figure it feels good to know
it works, and it only runs twice a week.

And -- since there's always something new -- this will be my chance to
play around with the `PATH+whatever=xx` functionality to prepend xx to
$PATH.

Test Plan:
I ran
   groovy jobs/sync-captions.groovy

But I'll try to run this as well once it's committed.

Reviewers: james

Reviewed By: james

Subscribers: colin, chris, amos

Differential Revision: https://phabricator.khanacademy.org/D33976