Groovy job for update-i18n-lite-videos.
Summary:
This one is really straightforward -- it just runs a shell script.

The only change from the old-style job is I also send a slack message
on job success: it's good to get some good news once ina while, and
this only runs twice a week.

However I still only send mail on failure and back-to-normal.

Test Plan:
I ran
    groovy jobs/update-i18n-lite-videos.groovy
But the real test will be running this for reals once it's landed.

Reviewers: james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D33977