Add support for notifying when a job is aborted.
In groovy-pipeline-land, jenkins doesn't try to modify the job status
until after the job has finished.  So if a job has been aborted,
there's no way to know that fact while the job is going on.  That
means, in particular, that our `notify` command doesn't know, when a
job fails, whether it's failed due to an abort request or not.

But there's a trick for figuring it out!  Someone put it on
stackexchange.  It's pretty clever, I thought.  Now notify uses it.

Auditors: amos, colin, chris

Test plan:
I put the new notify into a deploy-react-render-server job and ran it,
and aborted it in the middle, and saw the console say:
   ===== JOB ABORTED ====
and the deploy slack room said "deploy-react-render-server was aborted."

I then re-ran the deploy-react-render-server job again with the new
notify, but with an `error("hi")` after the `installDeps` step, and
saw it say
   ===== JOB FAILED ====
and the deploy slack room said "deploy-react-render-server failed."