Do not show partial results when we fail to get all the .pickle files.
Suppose we did not get all the pickle files back from the workers
(because one worker died before it could produce a pickle file,
perhaps due to an infrastructure bug).  One could imagine it's useful
to go on in this case, and analyze the pickle-file we *did* get back.
But in my experience it's too confusing: people think that the results
we emit are the full results, even though this error indicates some
results could not be processed.

Auditors: amos, colin, chris

Test Plan:
None