Modify dropbox to use a modern cert.
The cert file that comes with the super-old dropbox module we use
isn't up to date.  Use the more up-to-date system cert file.  Sadly,
this seems to be hard-coded into dropbox.py.

Auditors: james

Test plan:
I manually ran, on jenkins (as the jenkins user)
   cd /mnt/jenkins_jobs/i18n/jobs/sync-captions/workspace/webapp
   . ../env/bin/activate
   env PYTHONPATH=$HOME/secrets_py tools/dropbox_sync_source.py .
Before this change it failed with:
   dropbox.rest.RESTSocketError: Error connecting to "api.dropbox.com": [Errno 1] _ssl.c:510: error:14090086:SSL routines:SSL3_GET_SERVER_CERTIFICATE:certificate verify failed

After this change, it did not fail.