Add a convenience method, exec.outputOf.
Summary:
Oftentimes, I find myself doing
   sh(script: "...", returnStdout: true).trim();
(This is the equivalent of backticks in perl.)

It's verbose, so I wrote a shortcut for it.  I could have written a
short-cut for the `sh` version, but since I already have an
exec.groovy, I just did it for `exec` so I didn't have to create a new
library file.  Almost all the places where I want to use this, I can
use `exec` as easily as -- or sometimes more easily than -- `sh`.

Test Plan:
I ran
   groovy foo.groovy
for all the foo.groovy modified in this commit, and got no errors.

Reviewers: amos, colin, chris

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D34070