Get rid of helper scripts that are no longer used!
In the new groovy world, we do not need as many sh scripts as we did
before, because we can do a lot of control-flow stuff in groovy.  So
get rid of the stuff we don't need.

(analyze_make_output went away because we're finally using pickle-file
test-data everywhere to analyze test output.)

Auditors: amos, colin, chris

Test Plan:
I ran
   git grep -l -e analyze_make_output -e e2e-tests.sh -e build-make-target.sh
in both the `jenkins` and `jenkins-tools` repo (and `webapp` for good
measure).  I then looked at the matching files and made sure none
corresponded to currently running jobs (or to other scripts).