Move from the concurrent-builds plugin to the lock plugin.
Summary:
Previously, we were using the concurrent-builds plugin when we needed to
keep two jobs from running at the same time due to some global resource
contention (e.g. running-on-saucelabs).

Now, we use locks instead, via
https://wiki.jenkins-ci.org/display/JENKINS/Lockable+Resources+Plugin

The main advantage of locks is that they were designed for pipelines
from the beginning, which means that you can acquire a lock for a
period of time, rather than forcing it to be global for the entire
script.

We use that fact to get rid of the hack we were doing in
i18n-gcs-upload.groovy, where we depended on our parents to acquire a
concurrent-build lock that we needed, due to the global nature of
those locks (and the fact they aren't re-entrant).

We can also hold locks for less time now.  In particular, the
uses-a-lot-of-memory lock we only acquire when running a script that
actually uses a lot of memory.  For jobs that do work beyond what that
script does, this will allow for more parallelism than we had in the
past.

Test Plan:
Fingers crossed.  Will monitor some of these jobs as they run in the
future.

Reviewers: amos, chris, colin

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D34101