Use the new `--include-changes-since-data-filename` arg.
I made a breaking-API change to tests_for.py, to make it not include
quite so much history when looking for tests to run.  But we want all
that history at deploy-time, where we want to be really conservative
in what tests we run.  So I add the flag the brings back the old
behavior.

Auditors: benkraft

Test plan:
None