A groovy version of the deploy jenkins jobs.
Summary:
This script combines the work of deploy-via-multijob, deploy-default,
deploy-set-default, deploy-promote, deploy-monitor, and deploy-finish
into one happy script! This is the capstone of the "move to groovy"
project.

The script is long(ish), of course, but I think quite readable. I
tried to make the change as minimal as possible, in that we still use
deploy_pipeline.py for a lot of the logic even though it's a) doing a
lot of extra work we don't need now (like managing the deploy lock)
and b) harder to follow since the logic keeps jumping between the
groovy code and that script. But we can improve that over time.

I use `input()` for the places where we need to wait for user
input. Luckily, Jenkins expects that people might want to have
scripts/bots respond to input requests for you, so there's a simple
API for doing so. I don't expect it to be difficult to modify sun to
respond to input()s rather than its current behavior of starting new
jobs.

Got my logic backwards for DEPLOY_STATIC/DEPLOY_DYNAMIC.

Test Plan:
Even the longest journey begins with a single step.

(None new)

Reviewers: amos, chris, colin, benkraft

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D34183