Use the rc-output of should_deploy rather than the string output.
I wrote should_deploy to return 1 when we should deploy.  Jenkins was
interpreting that `1` as an error and aborting.  Probably my original
decision was pretty bone-headed, but rather than fix it I take
advantage of it for its intended purpose.

I wrote a new utility routine to help with that, since I use it in a
few places now.

Auditors: amos, colin, chris

Test plan:
Will try it!