First steps toward reducing the need for deploy_pipeline.py.
Summary:
deploy_pipeline.py is a bit overkill in the groovy world, since it
does a lot of orchestration that we now do in groovy.  In particular
we don't need to keep around a state file anymore.  So I'd like to
just remove our use of it.

But one place we depend on deploy_pipeline.py a lot is in sending
messages to slack.  So this code moves those slack messages into
groovy.  Since it's just data, that's not hard to do, though I had fun
writing the function that converted that data to the proper arguments
to alert.py.

I use this in one place so far: finish-with-success.  I moved the
logic of finish-with-success here too, since it's mostly just
executing shell commands.  I'm of two minds about this: it makes the
script bigger, and could easily be in a shell script of its own, but
on the other hand I hate the idea of having lots of different tiny
shell scripts that deploy-webapp.groovy calls out to.

In future commits, I'll work my way back toward acquire-lock.

Test Plan: groovy jobs/deploy-webapp.groovy

Reviewers: amos, colin, benkraft

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D34625