Remove the lockdir at the start of each job.
The lockdir is there in the old-style jenkins world because in that
case the deploy is separated across many jenkins jobs, and we wanted
to make sure they ran in the right order.  But in the groovy world
that is unnecessary, since everything is in a single job!

But we can't get rid of the lockdir just yet because we still call
into deploy_pipeline.py, which depends on it since it's still used for
old-style jobs.  So we still support it, just in a lackadaisical
fashion.  In particular, we're not good about cleaning up the lockdir
after we run, and we just delete it at script-start.

(Note that this lockdir is different from the one the old-style jobs
use, which lives in a different directory.  //This// lockdir never
actually does any locking -- the jenkins one-build-at-a-time rule does
that for us -- and is there only so deploy_pipeline.py doesn't
complain.)

Auditors: amos, colin

Test plan:
None