Use UTC when parsing translation dates.
We had a DST-related failure last weekend!
```
[main] ++ date -d '2017-03-12 02:12' +%s
[main] date: invalid date '2017-03-12 02:12'
```

This is because the 2am hour did not exist in pacific time.

This leads me to believe that these timestamps are UTC.  It's hard to
tell from the code -- we eventually get the timestamps from crowdin --
but for our purposes, being off by 8 hours or so doesn't matter
anyway.

Auditors: james

Test plan:
Ran this manually:
    % date -d '2017-03-12 02:12' +%s
    date: invalid date '2017-03-12 02:12'
    % env TZ=UTC date -d '2017-03-12 02:12' +%s
    1489284720

So with the `env` it works rather than failing.