Copy the MFMS prefill files around when copying current.sqlite
Summary:
This is a companion diff to D34408 and will be deployed at the same time.  The
main change is to copy around the files that that diff creates; no changes to
sync are necessary since it already does a publish.  In order to construct the
filenames correctly, I changed CURRENT_SQLITE_BUCKET to be a bucket instead of
a filename.  While I was at it I changed the other arg, SNAPSHOT_BUCKETS, to be
called SNAPSHOT_NAMES, since that's what it is now.

Test Plan:
- deploy it with D34408
- keep an eye on the next content sync
- after the next content sync, run `make current.sqlite` and see that it copies
  down the new files, and that they work right.

Reviewers: amos, csilvers

Differential Revision: https://phabricator.khanacademy.org/D34409