Do a better job of noticing aborts during monitoring.
I did a new-style deploy today, and aborted during monitoring.
Here is what jenkins said:
```
10:51:09 [monitor] Sending interrupt signal to process
[_watchdog] Failed in branch _watchdog
10:51:10 [monitor] Terminated
[Pipeline] [monitor] echo
10:51:10 [monitor] Marking unstable due to monitoring failure: hudson.AbortException: script returned exit code 143
```
and then it went on to the next step, same as if the monitoring
failure were due to not being able to talk to the error-monitor-db or
something.

But if you abort during monitoring, we want it to abort!  So now we
check for the explicitly.

Auditors: amos, colin

Test Plan:
Fingers crossed, but at least `groovy jobs/deploy-webapp.groovy` passes.