Inline the contents of deploy.sh.
Summary:
deploy.sh goes through all these hoops to run the deploy-to-gae and
deploy-to-gcs in parallel.  This is much easier done with the groovy
`parallel` functionality.  So now we do this.  The groovy code is
barely longer, and it's a whole lot more readable.

In the process, I made it so we properly do `make deps` and `clean` at
initialization-time.  By not having to run `make deps` again at deploy
time, we speed up the build a little bit!

Test Plan:
groovy jobs/deploy-webapp.groovy.

But I'll have to do a full deploy with this to make sure it works
right.

Reviewers: colin, amos, benkraft

Reviewed By: benkraft

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D34626