Make sure we always end up in `workspace`, not `workspace@2`.
This is a danger of running things in parallel: jenkins pipelines
wants to be helpful and give each job its own workspace.  Don't do
that jenkins!  But it does.

Last time I resolved this by putting the parallel *inside* the
`node()`.  But I can't do that here because one of the parallel things
we do is spawn another job, and it's not great to do that inside a
node (it keeps the executor tied up when it doesn't need to be).  So I
do a janky `cd` instead.

Auditors: amos, colin

Test plan:
Will try again