Do finishWithFailure() if we fail during deploying/testing.
I was having a bunch of deploy-time failures (due to scripting errors
on my part), and it was correctly saying "job failed" but not doing it
in the friendly way that alerts me due to the at-mention.

I realized this is because I wasn't asking it to.  So now I do.

It's a TODO to do this really early (before we acquire the
deploy-lock).  I'll be able to once we are fully done with using
deploy_pipeline.

Auditors: amos, colin

Test plan:
groovy jobs/deploy-webapp.groovy