Fix safe_sync_to_origin to *actually* sync to the origin.
I don't know how this was working before, and how webapp-maintenance
was successfully updating master.  But it seems like `git branch`
creates the branch but doesn't put you in it.  So we need to do that
explicitly as a second step.

The new `git checkout` doesn't actually change what commit you're at
-- `safe_sync_to` had already moved us to `$commit` -- it just changes
what the upstream is.

Auditors: dhruv

Test plan:
Fingers crossed -- we'll see if weekly-maintenance works next sat.