Do not include num-consecutive-failures when creating the asana task.
When `make allcheck` failed last night, it created this asana task:
   misc/make-allcheck failed (2nd time in a row)
This means that if it fails again night, it will want to create
   misc/make-allcheck failed (3rd time in a row)
which alertlib will think is a totally new failure, and will create a
new asana task for.

But we want alertlib to dedup these, and not create a new task every
night (the first task is enough until it's closed).  So now we just
say
   misc/make-allcheck failed

Auditors: amos, colin

Test plan:
   groovy vars/notify.groovy