Get rid of the need for build.lib in i18n-gcs-upload.groovy
Summary:
This is part of my effort to refactor build.lib.  Part of this is to
move a bunch of stuff that is done in shell scripts, into groovy code
instead.  This is the first step in that! -- a particularly easy one.

Test Plan: Fingers crossed -- will see if next i18n-gcs-upload job fails

Reviewers: james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D35117