Call build.lib as an app rather than importing it.
Summary:
This is part of my effort to refactor build.lib.  One way I'm doing
this is to treat it as a normal script rather than as a library you
include.

As part of this effort, I'm documenting requirements of where jenkins
scripts are called from.  (This should be fine since only the groovy
scripts call these shell scripts, and they can be in whatever
directory they need.)  That reduces our need to use bash, in favor of
normal sh.

Test Plan: Fingers crossed; next time this is run we'll see.

Reviewers: james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D35119