Oops, I need to redo all the build.lib commands now I'm in a new directory.
I'm now running the build.lib stuff from workspace-root rather than
from webapp-root, so all the arguments change.

This looks pretty ugly; it would be better to find another way of
doing it.  But it may be that the right solution is to move the
jenkins logic (and maybe this whole script!) into groovy.

Auditors: james

Test plan:
Fingers crossed