Call build.lib as an app rather than importing it.
Summary:
This is part of my effort to refactor build.lib.  One way I'm doing
this is to treat it as a normal script rather than as a library you
include.

While fixing this up, I noticed places I was calling `pwd` in the
wrong directory (I was supposed to do that //before// cd-ing to
webapp).  That should be fixed now.

Auditors: James

Test Plan:
Fingers crossed; next time this is run we'll see.