Call build.lib as an app rather than importing it.
Summary:
This is part of my effort to refactor build.lib.  One way I'm doing
this is to treat it as a normal script rather than as a library you
include.

While fixing this up, I noticed that it's not really correct to run
this in the webapp directory anymore, since it does lots of cleanup in
other places.  We shouldn't privilege webapp.

Test Plan:
Fingers crossed; we'll see how it does this weekend.  Maybe I'll run
it manually tonight.

Reviewers: dhruv, benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D35134