Call build.lib as an app rather than importing it.
Summary:
This is part of my effort to refactor build.lib.  One way I'm doing
this is to treat it as a normal script rather than as a library you
include.

sync-captions is a pretty large script but actually only uses
build.lib in one place.  So this was a pretty minimal change.

I'm tempted to rewrite sync-captions.sh in groovy, since the script
doesn't actually do all that much work.  But I'm resisting since what
we have now works just fine.

Get rid of an unnecessary `export`, based on review.

Test Plan:
Fingers crossed; next time this is run we'll see.

(same)

Reviewers: james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D35131