Jenkins job for khanalytics tests
Summary:
This commit sets up a simple job, modeled after webapp-test, for running the
khanalytics lint, flow checker, and unit tests.  At the moment, I have
installing the python deps and running the python unit tests disabled (with a
TODO to reenable) because I still need to figure out how to get python 3.6 and
a virtualenv running.  Lint also isn't working because it's trying to lint
python 2.  But before I muck around with that, I want to make sure
I've got the basics of a job working correctly.

Beyond the python version, there's also a ton of TODOs because the makefile
rules implicitly require docker, so we can't use them on jenkins.  In theory
they don't need to, but it's complicated to get that to work; for now I'm just
running the commands directly.  It's also still a TODO to do any sort of
reporting beyond success/failure.

Test Plan:
- `groovy khanalytics-test.groovy`, see that the only errors are missing
  imports
- create a job on jenkins and try it!
- see a success message in #future-of-pipelines

Reviewers: amos

Reviewed By: amos

Subscribers: tom, ragini, csilvers, #data-engineering

Differential Revision: https://phabricator.khanacademy.org/D35403