Call python3.6 directly for the linter
Summary:

The linter was failing (silently, since it was a linter problem and not an
actual linter error) since we were running a python2 linter on python3 code.
We've now got python3.6 installed on jenkins, so we'll call that directly to
start the linter.  This should correctly activate the python3 linter.  We
don't need to worry about setting up a virtualenv here because all the linter's
dependencies are vendored.

Test Plan:
- try it on Jenkins, verify that it's not failing due to string interpolation
  being a syntax error

Auditors: amos