Add `make deps` to deploy-znd
Summary:

This morning a number of znd deploys failed with error messages about missing
numpy or PIL or other deps.  After sshing into jenkins, it looked like the
issue was that there were no deps installed in the deploy-znd virtualenv, and
manually running `pip install -r requirements.txt` in the virtualenv fixed
things.

Looking at the scripts, it seems that the deploy-znd one doesn't run make deps,
whereas the deploy-webapp one does.  This commit adds a `make deps` step so
this doesn't happen again; I put it in the same place as for deploy-webapp,
right before the cleaning step.  My guess is that we didn't notice because the
virtualenv persists between deploys?  And someone must have recreated the
jenkins machine or otherwise done something that wipes out the virtualenvs in
the last week?

Test Plan:
- fingers crossed

Auditors: amos, csilvers