Also sync the linter as part of _setupWebapp
Summary:

See https://app.asana.com/0/41295607232218/333411946093389

The khan-linter repository is not auto-updating on jenkins workers.  Webapp
prefers to use the repo to run its linters (so that this is in sync with what
people run via `arc lint`).  It being out of date means that lint tests can
fail when `arc lint` passed.  This commit includes an update of the linter in
the `_setupWebapp` step so we always have the latest linter.

Test Plan:
- try it on jenkins
- ssh into a worker, see that ~/webapp-workspace/devtools/khan-linter is up to
  date

Auditors: amos, csilvers