Add GIT_REVISION param to build current.sqlite job.
I want to test some changes to the sync process, but they
are annoying to test locally since the job takes so long
to run. I would like to test them on jenkins. This change
to the job allows me to push my changes to a branch and have
the job use my branch rather than master.

Test Plan: None

Auditors: csilvers