Add a jenkins job to send the recent deploy history to slack
Summary:
See https://app.asana.com/0/31965416896056/344186384206728

Sometimes when we're tracking down a new error or problem, it's useful
to be able to quickly find everything that was deployed to production
recently.  This commit adds a jenkins job to send this information to
slack (which will eventually be run from a sun command).

Rather than wrestling with the jenkins API to find the last successful
deploy-webapp jobs, I instead opted to exploit the fact that we create
git tags for successful deploys when they're merged to master.  This
allows us to find the last 5 deploys by grabbing the most recent tags
(excluding any that were rolled back and marked with the `-bad` suffix).

To print the changelog, I reused the function for doing that that exists
in the webapp deploy package.

The only thing this strategy doesn't get us vs the actual jenkins jobs
is the name of the actual deployer, but I don't think this is a big deal
since we do print the names of all the committers, so there's always
someone to go contact.

Test Plan:
- create a pipeline job `deploy-history` on jenkins that pulls from a
  branch of jenkins-tools with this commit
- run the job (reporting to #bot-testing)
- verify it prints the five most recent deploys and their changelogs

Reviewers: csilvers, amos

Reviewed By: csilvers

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D35889