Move the path-munging to bigfile-specific callsites.
I removed the path-munging that manually set up the virtualenv.
Everyone should be setting up the virtualenv in groovy before calling
any build.lib routines anyway (`onMaster` and friends do this
automatically) so doing it here was duplicative.  Also, nothing in
build.lib actually //sets up// the virtualenv, so adding it to the
path might be overly optimistic anyway.  Now we have all our
virtualenv code only in one place, which is good.

The only path-munging we had left after that was for 'git bigfile'.
So I moved that into the git-bigfile specific callsites.  There were
two, which means there's some code duplication, but I think it's worth
it to minimize the amount of global state that build.lib modifies.
All we have left is the in-the-right-dir check -- which I hope to get
rid of too --- and then it's easy to start moving this functionality
into groovy as it makes sense.

Auditors: amos, colin

Test Plan:
Mostly fingers crossed, but I did do
    jenkins-tools/build.lib busy_wait_on_dropbox
and it failed on `dropbox.py: not found`, which means it read and
parsed build.lib no problem.